Intents between Popups

+++ app/src/main/java/com/example/multiboard/StrokePath.java
@@ -8,7 +8,7 @@ public class StrokePath {
     private int strokeWidth;
     private Path path;
 
-    public StrokePath(int color, int strokeWidth, Path path) {
+    public StrokePath(int color, float strokeWidth, Path path) {
         this.color = color;
         this.strokeWidth = strokeWidth;
         this.path = path;
